// api/apiMethods.types.ts
export interface paths {
    "/api/v1/auth/login": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * User login
         * @description Authenticates a user and returns a JWT access token.
         */
        post: operations["login"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/auth/register": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * User registration
         * @description Registers a new user
         */
        post: operations["register"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/auth/verify-email": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Verify user email
         * @description Verifies a user's email using an OTP code.
         */
        post: operations["verifyEmail"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/auth/resend-otp": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Resend OTP code
         * @description Resends the OTP code to the user's email address.
         */
        post: operations["resendOtp"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/users": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all users
         * @description Retrieves a list of all users in the system.
         */
        get: operations["getAllUsers"];
        put?: never;
        /**
         * Create a new user
         * @description Creates a new user in the system.
         */
        post: operations["createUser"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/users/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["UsersController_findOne"];
        put?: never;
        post?: never;
        delete: operations["UsersController_remove"];
        options?: never;
        head?: never;
        patch: operations["UsersController_update"];
        trace?: never;
    };
    "/api/v1/categories": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get all categories */
        get: operations["getAllCategories"];
        put?: never;
        /** Create category */
        post: operations["createCategory"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/categories/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["CategoriesController_findOne"];
        put?: never;
        post?: never;
        /** Delete category */
        delete: operations["deleteCategory"];
        options?: never;
        head?: never;
        patch: operations["CategoriesController_update"];
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        LoginDto: {
            /**
             * @description The username of the user
             * @example johndoe
             */
            username: string;
            /**
             * @description The password of the user
             * @example strongpassword123
             */
            password: string;
        };
        LoginResponseDto: {
            /**
             * @description JWT access token
             * @example eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImpvaG5kb2UiLCJpYXQiOjE2MTYyMzkwMjJ9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
             */
            accessToken: string;
        };
        RegisterDto: {
            /**
             * @description Username of the user
             * @example john_doe
             */
            username: string;
            /**
             * @description Password of the user
             * @example StrongP@ssw0rd!
             */
            password: string;
            /**
             * @description Email of the user
             * @example john_doe@example.com
             */
            email: string;
        };
        VerifyEmailDto: {
            /**
             * @description User email address
             * @example user@example.com
             */
            email: string;
            /**
             * @description OTP code sent to user email
             * @example 123456
             */
            otpCode: string;
        };
        ReSendEmailDto: {
            /**
             * @description User email address
             * @example user@example.com
             */
            email: string;
        };
        CreateUserDto: {
            /**
             * @description The username of the user
             * @example johndoe
             */
            username: string;
            /**
             * @description The password of the user
             * @example strongpassword123
             */
            password: string;
            /**
             * @description The email of the user
             * @example user@example.com
             */
            email: string;
        };
        UpdateUserDto: Record<string, never>;
        LanguageDto: {
            /** @default  */
            uz: string;
            /** @default  */
            ru: string;
            /** @default  */
            en: string;
            /** @default  */
            krill: string;
        };
        CreateCategoryDto: {
            name: components["schemas"]["LanguageDto"];
        };
        UpdateCategoryDto: Record<string, never>;
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    login: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["LoginDto"];
            };
        };
        responses: {
            default: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LoginResponseDto"];
                };
            };
        };
    };
// ...
    UsersController_findOne: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
    createCategory: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateCategoryDto"];
            };
        };
        responses: {
            /** @description The category has been successfully created. */
            201: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}